pyproject.toml
[project]
name = "odin"
version = "0.1.0"
description = "A modern, protocol-agnostic agent development framework with first-class support for MCP, A2A, and AG-UI"
authors = [
    {name = "Wu Zhipeng", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
keywords = ["agent", "ai", "mcp", "a2a", "ag-ui", "llm", "framework"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    # Core async runtime
    "anyio>=4.6.0",
    # Configuration management
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    # Structured logging
    "structlog>=24.4.0",
    "python-json-logger>=3.2.0",
    # OpenTelemetry for tracing and metrics
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp>=1.28.0",
    "opentelemetry-instrumentation>=0.49b0",
    "opentelemetry-exporter-prometheus>=0.49b0",
    "prometheus-client>=0.21.0",
    # Protocol support
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=14.0",
    "sse-starlette>=2.2.0",
    "starlette>=0.41.0",
    # MCP protocol
    "mcp>=1.3.1",
    # Data persistence
    "redis>=5.2.0",
    "aiosqlite>=0.20.0",
    # Utilities
    "tenacity>=9.0.0",  # Retry logic
    "click>=8.1.7",     # CLI framework
    "copilotkit>=0.1.72",
    "langgraph>=1.0.4",
    "langchain-openai>=1.1.0",
    "httpx[socks]>=0.28.1",
    "ag-ui-langgraph>=0.0.21",
]

[project.optional-dependencies]
# CopilotKit integration
copilotkit = [
    "copilotkit>=0.1.39",
]

# Development dependencies
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.13.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
]

# Performance profiling
profiling = [
    "pyinstrument>=5.0.0",
    "memory-profiler>=0.61.0",
    "py-spy>=0.4.0",
]

# Additional protocol support
protocols = [
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "protobuf>=5.29.0",
]

# PostgreSQL support
postgres = [
    "asyncpg>=0.30.0",
]

# NotebookLLM automation tools
notebookllm = [
    "playwright>=1.49.0",
    "pdf2image>=1.17.0",
]

# Built-in plugins dependencies
builtin-plugins = [
    "aiohttp>=3.9.0",
    "playwright>=1.49.0",
]

# PPT conversion tools (OCR + image processing)
pptx-converter = [
    "paddlepaddle>=3.0.0",
    "paddleocr>=3.0.0",  # Use latest; odin.compat patches langchain compatibility
    "python-pptx>=1.0.0",
    "opencv-python>=4.10.0",
    "langchain-core>=0.2.0",  # Required for paddleocr 3.x compatibility shim
    "langchain-text-splitters>=0.2.0",  # Required for paddleocr 3.x compatibility shim
    # Optional: better inpainting with LaMa
    # "lama-cleaner>=1.2.0",
]

# All optional dependencies
all = [
    "odin[dev,profiling,protocols,postgres,notebookllm,builtin-plugins,pptx-converter]",
]

[project.urls]
Homepage = "https://github.com/yourusername/odin"
Documentation = "https://odin.readthedocs.io"
Repository = "https://github.com/yourusername/odin"
Issues = "https://github.com/yourusername/odin/issues"

[project.scripts]
odin = "odin.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/odin"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=odin",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
]

[tool.coverage.run]
source = ["src/odin"]
branch = true
parallel = true
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod",
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
strict_concatenate = true

[[tool.mypy.overrides]]
module = [
    "structlog.*",
    "opentelemetry.*",
    "mcp.*",
    "copilotkit.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py314"
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "B008",   # Do not perform function calls in argument defaults
    "B905",   # `zip()` without an explicit `strict=` parameter
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"tests/*" = ["ARG", "S101"]  # Allow assert and unused arguments in tests

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.6",
]
